requirements.txt
openai
langchain
langchain-openai
python-dotenv
streamlit